Trait ra_parser::TokenSource

pub trait TokenSource {
    fn current(&self) -> Token;

    fn lookahead_nth(&self, n: usize) -> Token;

    fn bump(&mut self);

    fn is_keyword(&self, kw: &str) -> bool;
}

TokenSource abstracts the source of the tokens parser operates on.

Hopefully this will allow us to treat text and token trees in the same way!

Required methods

fn current(&self) -> Token

fn lookahead_nth(&self, n: usize) -> Token

Lookahead n token

fn bump(&mut self)

bump cursor to next token

fn is_keyword(&self, kw: &str) -> bool

Is the current token a specified keyword?

Loading content...