Crate ra_parser

The Rust parser.

The parser doesn't know about concrete representation of tokens and syntax trees. Abstract TokenSource and TreeSink traits are used instead. As a consequence, this crates does not contain a lexer.

The Parser struct from the parser module is a cursor into the sequence of tokens. Parsing routines use Parser to inspect current state and advance the parsing.

The actual parsing happens in the grammar module.

Tests for this crate live in ra_syntax crate.

Macros

T

token_set

Structs

ParseError

Reparser
A parsing function for a specific braced-block.

Token
TokenCursor abstracts the cursor of TokenSource operates one.

Enums

SyntaxKind
The kind of syntax node, e.g. IDENT, USE_KW, or STRUCT_DEF.

Traits

TokenSource
TokenSource abstracts the source of the tokens parser operates one.

TreeSink
TreeSink abstracts details of a particular syntax tree implementation.

Functions

parse
Parse given tokens into the given sink as a rust file.

parse_block
Parse given tokens into the given sink as a block

parse_expr
Parse given tokens into the given sink as a expression

parse_item
Parse given tokens into the given sink as an item

parse_macro_items

parse_macro_stmts

parse_meta

parse_pat
Parse given tokens into the given sink as a pattern

parse_path
Parse given tokens into the given sink as a path

parse_stmt
Parse given tokens into the given sink as a statement

parse_ty
Parse given tokens into the given sink as a ty

parse_vis
Parse given tokens into the given sink as an visibility qualifier